Clear architectural debt
  1. Created routers/responses/ module - New shared module similar to routers/conversations/
    - routers/responses/mod.rs - Module declaration
    - routers/responses/mcp_connection.rs - Contains ensure_request_mcp_client function
  2. Moved ensure_client_from_request_tools from mcp/manager.rs to routers/responses/mcp_connection.rs
    - Renamed to ensure_request_mcp_client as a free function
    - Removed the method from McpManager
    - Removed protocols::responses import from the mcp module (as you requested)
  3. Deleted routers/openai/conversations.rs - Was just re-exporting from the shared module, no longer needed
  4. Updated all callers:
    - routers/grpc/common/responses/utils.rs - Updated import path
    - routers/openai/router.rs - Updated import path
    - routers/openai/streaming.rs - Changed from method call to free function
